feat: update touchid functions

+++ plugins/functions.plugin.zsh
@@ -400,21 +400,29 @@ trm() {
 
 enable_touchid_sudo() {
   # Navigate to the directory containing the PAM configuration files
-  cd /etc/pam.d || return
+  pushd /etc/pam.d
+
+  if [[ -f "sudo_local" ]]; then
+    echo "sudo_local already exists. Touch ID for sudo is already enabled."
+    popd
+    return
+  fi
 
   # Copy the template file to create a new sudo_local file
   echo "Copying sudo_local.template to sudo_local. Please enter your sudo password if prompted."
   sudo cp sudo_local.template sudo_local
-  if [ $? -ne 0 ]; then
+  if [[ $? -ne 0 ]]; then
     echo "Failed to copy sudo_local.template. Ensure it exists and you have permissions."
+    popd
     return
   fi
 
   # Use sed to uncomment the line containing 'pam_tid.so'
   echo "Enabling Touch ID in sudo_local. You might need to enter your sudo password again."
   sudo sed -i '' 's/#\(.*pam_tid.so\)/\1/' sudo_local
-  if [ $? -ne 0 ]; then
+  if [[ $? -ne 0 ]]; then
     echo "Failed to enable Touch ID in sudo_local. Check your permissions and file content."
+    popd
     return
   fi
 
@@ -423,20 +431,22 @@ enable_touchid_sudo() {
 
 disable_touchid_sudo() {
   # Navigate to the directory containing the PAM configuration files
-  cd /etc/pam.d || return
+  pushd /etc/pam.d
 
   # Check if sudo_local exists before attempting to remove it
-  if [ -f "sudo_local" ]; then
+  if [[ -f "sudo_local" ]]; then
     echo "Removing sudo_local to revert to default sudo configuration. Please enter your sudo password if prompted."
     sudo rm sudo_local
     if [ $? -ne 0 ]; then
       echo "Failed to remove sudo_local. Ensure you have permissions."
+      popd
       return
     fi
     echo "sudo_local has been successfully removed. The system has reverted to the default sudo configuration."
   else
     echo "sudo_local does not exist. No changes needed."
   fi
+  popd
 }
 
 # select random element from arguments
